// frontend/src/types.ts
export type Store = {
  id?: number
  codigo_wsiga: string
  nit: string
  nombre_almacen: string
  direccion: string
  centro_comercial: string
  telefono_fijo: string
  celular: string
  whatsapp: string
  correo_electronico: string
  pagina_web: string
  facebook: string
  instagram: string
  tiktok: string
  created_at?: string
  updated_at?: string
}

export type User = {
  username: string
}